+++ src/resources/database/migrations/2025_01_13_092917_add_reference_to_the_customer_purchases_table.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+    public function up(): void
+    {
+        Schema::table('customer_purchases', function (Blueprint $table) {
+            $table->string('reference')->nullable()->after('purchase_id');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('customer_purchases', function (Blueprint $table) {
+            $table->dropColumn('reference');
+        });
+    }
+};