+++ src/resources/database/migrations/2025_01_13_094818_get_rid_of_smurf_naming_in_customer_purchases_table.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    public function up(): void
+    {
+        Schema::table('customer_purchases', function (Blueprint $table) {
+            $table->renameColumn('purchase_date', 'date');
+            $table->renameColumn('purchase_value', 'value');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('customer_purchases', function (Blueprint $table) {
+            $table->renameColumn('date', 'purchase_date');
+            $table->renameColumn('value', 'purchase_value');
+        });
+    }
+};