+++ tests/CustomerPurchaseTest.php
@@ -58,8 +58,8 @@ class CustomerPurchaseTest extends TestCase
 
         $customerPurchase = CustomerPurchaseProxy::create([
             'customer_id' => $customer->id,
-            'purchase_date' => '2024-11-21',
-            'purchase_value' => 89.62,
+            'date' => '2024-11-21',
+            'value' => 89.62,
             'currency' => 'EUR',
             'purchasable_id' => $order->id,
             'purchasable_type' => Order::class,
@@ -68,9 +68,9 @@ class CustomerPurchaseTest extends TestCase
 
         $this->assertTrue($order->is($customerPurchase->purchasable));
         $this->assertTrue($customer->is($customerPurchase->customer));
-        $this->assertEquals(89.62, $customerPurchase->purchase_value);
+        $this->assertEquals(89.62, $customerPurchase->value);
         $this->assertEquals('EUR', $customerPurchase->currency);
-        $this->assertEquals('2024-11-21', $customerPurchase->purchase_date->toDateString());
+        $this->assertEquals('2024-11-21', $customerPurchase->date->toDateString());
         $this->assertEquals('70H-0WJB-2OD6', $customerPurchase->reference);
     }
 
@@ -93,8 +93,8 @@ class CustomerPurchaseTest extends TestCase
 
         $customerPurchase = CustomerPurchaseProxy::create([
             'customer_id' => $customer->id,
-            'purchase_date' => '2024-11-21',
-            'purchase_value' => 89.62,
+            'date' => '2024-11-21',
+            'value' => 89.62,
             'currency' => 'EUR',
             'purchasable_id' => $order->id,
             'purchasable_type' => Order::class,
@@ -102,8 +102,8 @@ class CustomerPurchaseTest extends TestCase
 
         $customerPurchase->update([
             'customer_id' => $customer2->id,
-            'purchase_date' => '2001-01-06',
-            'purchase_value' => 2001.06,
+            'date' => '2001-01-06',
+            'value' => 2001.06,
             'currency' => 'HUF',
             'purchasable_id' => $order2->id,
             'purchasable_type' => Order::class,
@@ -117,9 +117,9 @@ class CustomerPurchaseTest extends TestCase
         $this->assertFalse($customer->is($updatedCustomerPurchase->customer));
         $this->assertTrue($customer2->is($updatedCustomerPurchase->customer));
 
-        $this->assertEquals(2001.06, $updatedCustomerPurchase->purchase_value);
+        $this->assertEquals(2001.06, $updatedCustomerPurchase->value);
         $this->assertEquals('HUF', $updatedCustomerPurchase->currency);
-        $this->assertEquals('2001-01-06', $updatedCustomerPurchase->purchase_date->toDateString());
+        $this->assertEquals('2001-01-06', $updatedCustomerPurchase->date->toDateString());
     }
 
     /**
@@ -135,8 +135,8 @@ class CustomerPurchaseTest extends TestCase
 
         $customerPurchase = CustomerPurchaseProxy::create([
             'customer_id' => $customer->id,
-            'purchase_date' => '2024-11-21',
-            'purchase_value' => 89.62,
+            'date' => '2024-11-21',
+            'value' => 89.62,
             'currency' => 'EUR',
             'purchasable_id' => $order->id,
             'purchasable_type' => Order::class,
@@ -166,8 +166,8 @@ class CustomerPurchaseTest extends TestCase
 
         $customerPurchase1 = CustomerPurchaseProxy::create([
             'customer_id' => $customer->id,
-            'purchase_date' => '2024-11-21',
-            'purchase_value' => 89.62,
+            'date' => '2024-11-21',
+            'value' => 89.62,
             'currency' => 'EUR',
             'purchasable_id' => $order->id,
             'purchasable_type' => Order::class,
@@ -175,8 +175,8 @@ class CustomerPurchaseTest extends TestCase
 
         $customerPurchase2 = CustomerPurchaseProxy::create([
             'customer_id' => $customer->id,
-            'purchase_date' => '2023-12-09',
-            'purchase_value' => 10000000.94,
+            'date' => '2023-12-09',
+            'value' => 10000000.94,
             'currency' => 'HUF',
             'purchasable_id' => $order2->id,
             'purchasable_type' => Order::class,
@@ -192,8 +192,8 @@ class CustomerPurchaseTest extends TestCase
             return $order instanceof Order;
         }));
 
-        $this->assertTrue($customer->purchases->pluck('purchase_value')->contains(89.62));
-        $this->assertTrue($customer->purchases->pluck('purchase_value')->contains(10000000.94));
+        $this->assertTrue($customer->purchases->pluck('value')->contains(89.62));
+        $this->assertTrue($customer->purchases->pluck('value')->contains(10000000.94));
 
         $this->assertTrue($customer->purchases->map->purchasable->pluck('number')->contains('50M3-0RD3R-NUMB3R'));
         $this->assertTrue($customer->purchases->map->purchasable->pluck('number')->contains('WH47-H4V3-Y0U-B0U6H7'));